Add support for appendices
This was quite a lot of work to be honest and it isn't done yet. To
really be guideline-compliant we still have to remove the dotted line
and the page numbers from the list of appendices, both of which I have
no idea on how to achieve it. Any help is welcome.