Add 4th parameter to \abb for another caption line
CAUTION: this commit breaks previous papers because it adds another
*mandatory* parameter to the \abb command. This is due to limitations
of the \newcommand command (you can only have one optional arguments).
However, the new mandatory command can be left empty. It is used to
display a second caption line which will not appear in the table of
figures which can be used for providing a source for a figure, for
example.